fix : change valuable type of motor[4] float to int16_t

+++ lib/raspSerial/raspSerial.cpp
@@ -17,10 +17,10 @@ void raspSerial::receiveRx() {
             bufferCount = 0;
             //受信バッファに溜まったデータを実用の変数に代入していく
             // for使うより直接代入の方が早い
-            info.motor[0] = (float)buffer[1] - 100.0;
-            info.motor[1] = (float)buffer[2] - 100.0;
-            info.motor[2] = (float)buffer[3] - 100.0;
-            info.motor[3] = (float)buffer[4] - 100.0;
+            info.motor[0] = (int16_t)buffer[1] - 100.0;
+            info.motor[1] = (int16_t)buffer[2] - 100.0;
+            info.motor[2] = (int16_t)buffer[3] - 100.0;
+            info.motor[3] = (int16_t)buffer[4] - 100.0;
             info.driblePower = (float)buffer[5] / 100;
             info.kickerPower = (float)buffer[6] / 100;
         }
